requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
crypto==1.4.1
cryptography==2.3.1
enum34==1.1.6
Flask==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.0
Naked==0.1.31
pycparser==2.19
pycryptodome==3.7.2
pycryptodomex==3.7.2
PyMySQL==0.9.2
PyPDF2==1.26.0
python-alipay-sdk==1.9.0
PyYAML==3.13
requests
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.23
weixin-python==0.5.4
Werkzeug==0.14.1
